feat: ajudst apis
