+++ examples/miniprogram-taro/package.json
@@ -36,8 +36,8 @@
   "author": "",
   "dependencies": {
     "@authing/miniprogram-jsencrypt": "^5.0.0-alpha.14",
-    "@authing/miniprogram-sm2encrypt": "^5.0.0-alpha.15",
-    "@authing/miniprogram-taro": "^5.0.0-alpha.17",
+    "@authing/miniprogram-sm2encrypt": "^5.0.0-alpha.20",
+    "@authing/miniprogram-taro": "^5.0.0-alpha.20",
     "@babel/runtime": "^7.7.7",
     "@tarojs/components": "3.5.4",
     "@tarojs/helper": "3.5.4",
