+++ examples/miniprogram-wx/miniprogram/miniprogram_npm/@authing/miniprogram-jsencrypt/index.js
@@ -4,7 +4,7 @@ var __DEFINE__ = function(modId, func, req) { var m = { exports: {}, _tempexport
 var __REQUIRE__ = function(modId, source) { if(!__MODS__[modId]) return require(source); if(!__MODS__[modId].status) { var m = __MODS__[modId].m; m._exports = m._tempexports; var desp = Object.getOwnPropertyDescriptor(m, "exports"); if (desp && desp.configurable) Object.defineProperty(m, "exports", { set: function (val) { if(typeof val === "object" && val !== m._exports) { m._exports.__proto__ = val.__proto__; Object.keys(val).forEach(function (k) { m._exports[k] = val[k]; }); } m._tempexports = val }, get: function () { return m._tempexports; } }); __MODS__[modId].status = 1; __MODS__[modId].func(__MODS__[modId].req, m, m.exports); } return __MODS__[modId].m.exports; };
 var __REQUIRE_WILDCARD__ = function(obj) { if(obj && obj.__esModule) { return obj; } else { var newObj = {}; if(obj != null) { for(var k in obj) { if (Object.prototype.hasOwnProperty.call(obj, k)) newObj[k] = obj[k]; } } newObj.default = obj; return newObj; } };
 var __REQUIRE_DEFAULT__ = function(obj) { return obj && obj.__esModule ? obj.default : obj; };
-__DEFINE__(1662259787368, function(require, module, exports) {
+__DEFINE__(1662259787377, function(require, module, exports) {
 var __TEMP__ = require('./jsencrypt');var JSEncrypt = __TEMP__['JSEncrypt'];
 
 if (!exports.__esModule) Object.defineProperty(exports, "__esModule", { value: true });function encryptFunction (plainText, publicKey) {
@@ -17,8 +17,8 @@ if (!exports.__esModule) Object.defineProperty(exports, "__esModule", { value: t
   return encStr.toString()
 };exports.encryptFunction = encryptFunction
 
-}, function(modId) {var map = {"./jsencrypt":1662259787369}; return __REQUIRE__(map[modId], modId); })
-__DEFINE__(1662259787369, function(require, module, exports) {
+}, function(modId) {var map = {"./jsencrypt":1662259787378}; return __REQUIRE__(map[modId], modId); })
+__DEFINE__(1662259787378, function(require, module, exports) {
 var window = {}
   ; (function (global, factory) {
     typeof exports === "object" && typeof module !== "undefined"
@@ -5645,7 +5645,7 @@ var window = {}
     Object.defineProperty(exports, "__esModule", { value: true })
   })
 }, function(modId) { var map = {}; return __REQUIRE__(map[modId], modId); })
-return __REQUIRE__(1662259787368);
+return __REQUIRE__(1662259787377);
 })()
 //miniprogram-npm-outsideDeps=[]
 //# sourceMappingURL=index.js.map
