+++ examples/miniprogram-wx/miniprogram/pages/index/index.js
@@ -23,7 +23,7 @@ Page({
 
     const { code } = await wx.login()
     
-    const res = await authing.core.loginByCode({
+    const res = await authing.loginByCode({
       connection: 'wechat_mini_program_code',
       extIdpConnidentifier: 'authing-zhaoyiming-miniprogram',
       wechatMiniProgramCodePayload: {
@@ -36,7 +36,7 @@ Page({
       }
     })
 
-    console.log('authing.core.loginByCode res: ', res)
+    console.log('authing.loginByCode res: ', res)
   },
 
   async loginByPhone (e) {
@@ -46,7 +46,7 @@ Page({
 
     const { code } = await wx.login()
     
-    const res = await authing.core.loginByPhone({
+    const res = await authing.loginByPhone({
       connection: 'wechat_mini_program_phone',
       extIdpConnidentifier: 'authing-zhaoyiming-miniprogram',
       wechatMiniProgramPhonePayload: {
@@ -59,7 +59,7 @@ Page({
       }
     })
 
-    console.log('authing.core.loginByPhone res: ', res)
+    console.log('authing.loginByPhone res: ', res)
   },
 
   /**
@@ -69,16 +69,16 @@ Page({
   async getPhone (e) {
     const { code } = e.detail
 
-    const res = await authing.user.getPhone({
+    const res = await authing.getPhone({
       extIdpConnidentifier: 'authing-zhaoyiming-miniprogram',
       code: '13a6011af350ce43bd8331d6af25f7dad596e273e41b66058b4d070f06ad33b2'
     })
 
-    console.log('authing.user.getPhone res: ', res)
+    console.log('authing.getPhone res: ', res)
   },
 
   async loginByPassword () {
-    const res = await authing.core.loginByPassword({
+    const res = await authing.loginByPassword({
       connection: 'PASSWORD',
       passwordPayload: {
         password: '123',
@@ -90,22 +90,22 @@ Page({
       }
     })
 
-    console.log('authing.core.loginByPassword res: ', res)
+    console.log('authing.loginByPassword res: ', res)
   },
 
   async sendSms () {
     // 指定 channel 为 CHANNEL_LOGIN，发送登录所用的验证码
-    const res = await authing.core.sendSms({
+    const res = await authing.sendSms({
       phoneNumber: '13100000000',
       phoneCountryCode: '+86',
       channel: 'CHANNEL_LOGIN'
     })
 
-    console.log('authing.core.sendSms res: ', res)
+    console.log('authing.sendSms res: ', res)
   },
 
   async loginByPassCode () {
-    const res = await authing.core.loginByPassCode({
+    const res = await authing.loginByPassCode({
       connection: 'PASSCODE',
       passCodePayload: {
         // 手机收到的短信验证码
@@ -118,39 +118,39 @@ Page({
       }
     })
 
-    console.log('authing.core.loginByPassCode: ', res)
+    console.log('authing.loginByPassCode: ', res)
   },
 
   async refreshToken () {
-    const res = await authing.core.refreshToken()
-    console.log('authing.core.refreshToken res: ', res)
+    const res = await authing.refreshToken()
+    console.log('authing.refreshToken res: ', res)
   },
 
   async updatePassword () {
-    const res = await authing.user.updatePassword({
+    const res = await authing.updatePassword({
       newPassword: '123',
       oldPassword: '123',
       passwordEncryptType: 'none'
     })
 
-    console.log('authing.user.updatePassword res: ', res)
+    console.log('authing.updatePassword res: ', res)
   },
 
   async getUserInfo () {
-    const res = await authing.user.getUserInfo()
-    console.log('authing.user.getUserInfo res: ', res)
+    const res = await authing.getUserInfo()
+    console.log('authing.getUserInfo res: ', res)
   },
 
   async updateAvatar () {
-    const res = await authing.user.updateAvatar()
-    console.log('authing.user.updateAvatar res: ', res)
+    const res = await authing.updateAvatar()
+    console.log('authing.updateAvatar res: ', res)
   },
 
   async updateUserInfo () {
-    const res = await authing.user.updateUserInfo({
+    const res = await authing.updateUserInfo({
       address: 'Hello world'
     })
 
-    console.log('authing.user.updateUserInfo res: ', res)
+    console.log('authing.updateUserInfo res: ', res)
   }
 })
